+++ src/app/survey/page.tsx
@@ -1,6 +1,13 @@
 import Survey from "@/components/Survey";
+import { isThereServerSession } from "@/hooks/session";
+import { redirect } from "next/navigation";
 
 const SurveyPage = async () => {
+  // const { session } = await isThereServerSession();
+  // if (!session) {
+  //   return redirect("/login");
+  // }
+
   return (
     <div>
       SurveyPage
