+++ src/hooks/session.ts
@@ -1,3 +1,4 @@
+import { User } from "@/lib/zustandStore";
 import {
   // createClientComponentClient,
   createServerComponentClient,
@@ -9,7 +10,7 @@ export const isThereServerSession = async () => {
   const {
     data: { session },
   } = await supabase.auth.getSession();
-  return { supabase, session };
+  return { session };
 };
 
 // export const isThereClientSession = async () => {
@@ -19,3 +20,15 @@ export const isThereServerSession = async () => {
 //   } = await supabase.auth.getSession();
 //   return { supabase, session };
 // };
+
+export const addUserReault = async (userResult: User) => {
+  "use server";
+  // await fetch("http://localhost:3000/user/survey", {
+  //     method: "post",
+  //     body: JSON.stringify({userResult})
+  // })
+  // router.refresh()
+  const supabase = createServerComponentClient({ cookies });
+  const { data, error } = await supabase.from("survey").insert(userResult);
+  return { data, error };
+};
