+++ src/middleware.ts
@@ -1,0 +1,29 @@
+import { createMiddlewareClient } from "@supabase/auth-helpers-nextjs";
+import { NextResponse } from "next/server";
+
+import type { NextRequest } from "next/server";
+
+export async function middleware(req: NextRequest) {
+  const res = NextResponse.next();
+
+  const supabase = createMiddlewareClient({ req, res });
+
+  const { data } = await supabase.auth.getSession();
+  if (!data.session && !req.nextUrl.pathname.startsWith("/login")) {
+    return NextResponse.redirect(new URL("/login", req.url));
+  } else {
+    return res;
+  }
+}
+
+export const config = {
+  matcher: [
+    /*
+     * Match all request paths "except for" the ones starting with:
+     * - _next/static (static files)
+     * - _next/image (image optimization files)
+     * - favicon.ico (favicon file)
+     */
+    "/((?!_next/static|_next/image|favicon.ico).*)",
+  ],
+};
